// FrontSPA/src/app/core/models/Entities/Patient.ts
export interface Patient {
    labCode?: string;
    degreeId?: string;
    degreeName?: string;
    degreeNameAr?: string;
    departmentId?: string;
    departmentName?: string;
    departmentNameAr?: string;
    name?: string;
    selected?: boolean;
    phone?: string;
    email?: string;
    dateOfBirth?: Date;
    nationalId?: string;
    state?: string;
    city?: string;
    street?: string;
    ageNow?: string;
    sex?: string
    id: string;
    lastEditDate?: string;
    createdDate?: string;

}